Enable running the nearest test within python
This makes a few changes to how the test scaffold works, but I think
overall they're not bad.  We begin with assuming that the argument to
a python test runner (nose, manage.py test, etc.) wants a python
module name.  This simplifies any script that's going to receive this
code, since most python things don't work at the file level, they want
that module spec.

We then give allow an optional command suffix.  This will be useful in
either ruby or python, though right now it's not used on the ruby
side of things.  For python tests, this suffix can be the class and
function name of the individual test to be run.  Not sure what key to
bind this to, yet.