Replace headlong/bookmarks with project.el
Project.el is built-in to emacs, but we're installing the Gnu ELPA
package to get features from emacs 28.

It's not quite the same experience, but both of these pieces of code
are designed for the same purpose.  Still need some fine tuning to
preserve the speed of the old technique, while also having the
advantages of the newer one (immediately activating the customizable
command menu, and then subsequently having access to the `C-x p` set
of commands.